+++ squirrel.frontier/src/test/java/org/dice_research/squirrel/queue/domainbased/MongoDBDomainQueueTest.java
@@ -13,11 +13,21 @@ import java.util.Collections;
 import java.util.Iterator;
 import java.util.List;
 
+import org.aksw.jena_sparql_api.core.QueryExecutionFactory;
+import org.aksw.jena_sparql_api.core.QueryExecutionFactoryDataset;
+import org.apache.jena.graph.Node;
+import org.apache.jena.graph.NodeFactory;
+import org.apache.jena.query.Dataset;
+import org.apache.jena.query.DatasetFactory;
+import org.apache.jena.rdf.model.ModelFactory;
+import org.apache.jena.sparql.core.DatasetGraph;
+import org.dice_research.squirrel.Constants;
 import org.dice_research.squirrel.MongoDBBasedTest;
 import org.dice_research.squirrel.data.uri.CrawleableUri;
 import org.dice_research.squirrel.data.uri.CrawleableUriFactory4Tests;
 import org.dice_research.squirrel.data.uri.UriType;
 import org.dice_research.squirrel.data.uri.UriUtils;
+import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
@@ -27,22 +37,62 @@ public class MongoDBDomainQueueTest extends MongoDBBasedTest {
     private List<CrawleableUri> uris = new ArrayList<CrawleableUri>();
     private String expectedDomains[];
     private MongoDBDomainBasedQueue mongodbQueue;
+    private QueryExecutionFactory queryExecFactory;
 
     @Before
     public void setUp() throws Exception {
-        mongodbQueue = new MongoDBDomainBasedQueue("localhost", 58027,false);
-
+        queryExecFactory = initQueryFactoryEngine();
+        mongodbQueue = new MongoDBDomainBasedQueue("localhost", 58027,false,queryExecFactory);
         CrawleableUriFactory4Tests cuf = new CrawleableUriFactory4Tests();
         uris.add(cuf.create(new URI("http://localhost/sparql"), InetAddress.getByName("127.0.0.1"), UriType.SPARQL));
-        uris.add(cuf.create(new URI("http://dbpedia.org/resource/New_York"), InetAddress.getByName("127.0.0.1"),
+        uris.add(cuf.create(new URI("http://dbpedia.org/resource/New_York_City"), InetAddress.getByName("127.0.0.1"),
                 UriType.DEREFERENCEABLE));
         uris.add(cuf.create(new URI("http://dbpedia.org/resource/Moscow"), InetAddress.getByName("127.0.0.1"),
                 UriType.DEREFERENCEABLE));
+        uris.add(cuf.create(new URI("http://dbpedia.org/resource/Berlin"), InetAddress.getByName("127.0.0.1"),
+            UriType.DEREFERENCEABLE));
+        uris.add(cuf.create(new URI("http://dbpedia.org/resource/Bangalore"), InetAddress.getByName("127.0.0.1"),
+            UriType.DEREFERENCEABLE));
         // Added this to check https://github.com/AKSW/Squirrel/issues/17
         uris.add(cuf.create(new URI("http://danbri.org/foaf.rdf"),
                 InetAddress.getByName((new URI("http://danbri.org/foaf.rdf")).toURL().getHost()), UriType.DUMP));
-        
+
         expectedDomains = new String[] {"danbri.org", "dbpedia.org", "localhost"};
+    }
+
+    private QueryExecutionFactory initQueryFactoryEngine() {
+        Dataset dataset = DatasetFactory.create();
+        dataset.setDefaultModel(ModelFactory.createDefaultModel());
+        Node g = NodeFactory.createURI(Constants.DEFAULT_META_DATA_GRAPH_URI.toString());
+//        DatasetGraph graph = createDataSetGraph(dataset, g);
+
+        Node s1 = NodeFactory.createURI("http://dbpedia.org/resource/New_York_City");
+        Node s2 = NodeFactory.createURI("http://dbpedia.org/resource/Berlin");
+        Node s3 = NodeFactory.createURI("http://dbpedia.org/resource/Bangalore");
+        Node s4 = NodeFactory.createURI("http://dbpedia.org/resource/Moscow");
+        String otherUri = "http://dbpedia.org/doesntMatter";
+        Node o = NodeFactory.createURI("http://dbpedia.org/doesntMatter");
+        DatasetGraph graph = dataset.asDatasetGraph();
+
+        for(int i = 0; i <= 5; i++) {
+            Node p = NodeFactory.createURI(otherUri + i);
+            graph.add(g, s1, p, o); // <http://dbpedia.org/resource/New_York_City>
+        }
+        for(int i = 0; i <= 10; i++) {
+            Node p = NodeFactory.createURI(otherUri + i);
+            graph.add(g, s2, p, o); // <http://dbpedia.org/resource/Berlin>
+        }
+        for(int i = 0; i <= 8; i++) {
+            Node p = NodeFactory.createURI(otherUri + i);
+            graph.add(g, s3, p, o); // <http://dbpedia.org/resource/Bangalore>
+        }
+        for(int i = 0; i <= 4; i++) {
+            Node p = NodeFactory.createURI(otherUri + i);
+            graph.add(g, s4, p, o); // <http://dbpedia.org/resource/Moscow>
+        }
+
+
+        return new QueryExecutionFactoryDataset(dataset);
     }
 
     @Test
@@ -98,6 +148,30 @@ public class MongoDBDomainQueueTest extends MongoDBBasedTest {
     }
 
     @Test
+    public void addCrawleableUriWithScore() throws Exception {
+        mongodbQueue.open();
+        mongodbQueue.purge();
+        for (CrawleableUri uri : uris) {
+            mongodbQueue.addUri(uri);
+        }
+        assertEquals(3, mongodbQueue.length());
+        List<CrawleableUri> listUris = mongodbQueue.getNextUris();
+        assertEquals(1, listUris.size());
+        List<CrawleableUri> listUris2 = mongodbQueue.getNextUris();
+        assertEquals(4, listUris2.size());
+        Assert.assertEquals("http://dbpedia.org/resource/Moscow", listUris2.get(0).getUri().toString());
+        Assert.assertEquals("http://dbpedia.org/resource/New_York_City", listUris2.get(1).getUri().toString());
+        Assert.assertEquals("http://dbpedia.org/resource/Bangalore", listUris2.get(2).getUri().toString());
+        Assert.assertEquals("http://dbpedia.org/resource/Berlin", listUris2.get(3).getUri().toString());
+        List<CrawleableUri> listUris3 = mongodbQueue.getNextUris();
+        assertEquals(1, listUris3.size());
+        List<CrawleableUri> listUris4 = mongodbQueue.getNextUris();
+        assertEquals(null, listUris4);
+        mongodbQueue.purge();
+        mongodbQueue.close();
+    }
+
+    @Test
     public void getIterator() throws Exception {
         mongodbQueue.open();
         mongodbQueue.purge();
@@ -121,7 +195,7 @@ public class MongoDBDomainQueueTest extends MongoDBBasedTest {
         for (CrawleableUri uri : uris) {
             mongodbQueue.addUri(uri);
         }
-        
+
         List<CrawleableUri> listUris = mongodbQueue.getNextUris();
         Iterator<String> iter = mongodbQueue.getGroupIterator();
         int count = 0;
@@ -133,7 +207,7 @@ public class MongoDBDomainQueueTest extends MongoDBBasedTest {
                 ++count;
             }
         }
-        
+
         assertEquals(uris.size(), count);
         mongodbQueue.close();
     }
@@ -147,7 +221,7 @@ public class MongoDBDomainQueueTest extends MongoDBBasedTest {
         assertEquals(1, retrievedUris.size());
         assertTrue(uris.get(1).equals(retrievedUris.get(0)));
         mongodbQueue.addUri(uris.get(2));
-        // The retrieval should fail 
+        // The retrieval should fail
         assertNull(mongodbQueue.getNextUris());
         // Give back the first URI
         mongodbQueue.markUrisAsAccessible(retrievedUris);
